feat: Complete comprehensive project restructure and refactoring
- Reorganize components into logical sub-folders (ui, habit, mood, charts, garden)
- Extract constants (colors, dimensions, settings) into separate files
- Separate styles from components into dedicated style files
- Rename explore.tsx to analytics.tsx for better naming
- Refactor large files into smaller, maintainable components
- Create reusable screen components for analytics and settings sections
- Fix all i18n translation keys and missing entries
- Clean up TypeScript and lint warnings
- Update Expo packages to compatible versions

File reductions:
- settings.tsx: 550 lines → 29 lines (95% reduction)
- analytics.tsx: 1289 lines → 634 lines (51% reduction)
