+++ components/charts/index.ts
@@ -1,0 +1,1 @@
+export { ProgressChart } from './ProgressChart';
