+++ components/garden/GardenGrowth.tsx
