+++ components/garden/index.ts
@@ -1,0 +1,1 @@
+export { GardenGrowth } from './GardenGrowth';
