+++ components/habit/AnimatedHabitItem.tsx
