+++ components/habit/index.ts
@@ -1,0 +1,3 @@
+export { AnimatedHabitItem } from './AnimatedHabitItem';
+export { HabitContextMenu } from './HabitContextMenu';
+export { RestartDialog } from './RestartDialog';
