+++ components/index.ts
@@ -1,0 +1,6 @@
+// Main components index - exports all components
+export * from './ui';
+export * from './habit';
+export * from './mood';
+export * from './charts';
+export * from './garden';
