+++ components/mood/MoodSelector.tsx
