+++ components/ui/haptic-tab.tsx
