+++ components/ui/index.ts
@@ -1,0 +1,2 @@
+export { ThemedCard } from './ThemedCard';
+export { HapticTab } from './haptic-tab';
