+++ screens/settings/index.ts
@@ -1,0 +1,6 @@
+// Export all settings screen components
+export { SettingsHeader } from './SettingsHeader';
+export { ThemeSection } from './ThemeSection';
+export { LanguageSection } from './LanguageSection';
+export { NotificationSection } from './NotificationSection';
+export { ExportSection } from './ExportSection';
